Move TestGetConnectionName to its own file
